+++ src/Handler/CurlMultiHandler.php
@@ -70,14 +70,7 @@ class CurlMultiHandler
     {
         $this->factory = $options['handle_factory'] ?? new CurlFactory(50);
 
-        if (isset($options['select_timeout'])) {
-            $this->selectTimeout = $options['select_timeout'];
-        } elseif ($selectTimeout = Utils::getenv('GUZZLE_CURL_SELECT_TIMEOUT')) {
-            trigger_deprecation('guzzlehttp/guzzle', '7.2.0', 'Using environment variable GUZZLE_CURL_SELECT_TIMEOUT is deprecated. Use option "select_timeout" instead.');
-            $this->selectTimeout = (int) $selectTimeout;
-        } else {
-            $this->selectTimeout = 1;
-        }
+        $this->selectTimeout = $options['select_timeout'] ?? 1;
 
         $this->options = $options['options'] ?? [];
 
@@ -231,12 +224,8 @@ class CurlMultiHandler
      *
      * @return bool True on success, false on failure.
      */
-    private function cancel($id): bool
+    private function cancel(int $id): bool
     {
-        if (!is_int($id)) {
-            trigger_deprecation('guzzlehttp/guzzle', '7.4', 'Not passing an integer to %s::%s() is deprecated and will cause an error in 8.0.', __CLASS__, __FUNCTION__);
-        }
-
         // Cannot cancel if it has been processed.
         if (!isset($this->handles[$id])) {
             return false;